Move shell from apps.c because of string.h issue
